test(mitm-addon): split usage buffer tests
